8th commit
エラーページ作成

+++ Diary/app/Http/Controllers/DiaryController.php
@@ -36,25 +36,42 @@ class DiaryController extends Controller
         return redirect()->route('diary.index');
     }
 
-    public function destroy(int $id)
+    public function destroy(Diary $diary)
     {
+        if (Auth::user()->id !== $diary->user_id) {
+            abort(403);
+        } 
         //Diaryモデルを使用して、diariesテーブルから$idと一致するidをもつデータを取得
-        $diary = Diary::find($id); 
+        // $diary = Diary::find($id); 
         //取得したデータを削除
         $diary->delete();
         return redirect()->route('diary.index');
     }
 
-    public function edit(int $id)
+    // public function edit({})
+    // {
+    //     $diary = Diary::find($id); 
+    //     return view('diaries.edit', [
+    //             'diary' => $diary,
+    //     ]);
+    // }
+    public function edit(Diary $diary)
     {
-        $diary = Diary::find($id); 
+        if (Auth::user()->id !== $diary->user_id) {
+            abort(403);
+        } 
+
         return view('diaries.edit', [
-                'diary' => $diary,
+            'diary' => $diary,
         ]);
     }
 
     public function update(int $id, CreateDiary $request)
     {
+        if (Auth::user()->id !== $diary->user_id) {
+            abort(403);
+        } 
+        
         $diary = Diary::find($id);
         $diary->title = $request->title; //画面で入力されたタイトルを代入
         $diary->body = $request->body; //画面で入力された本文を代入
